+++ packages/eslint-plugin-formatjs/tests/no-multiple-whitespaces.test.ts
@@ -99,5 +99,17 @@ ruleTester.run('no-multiple-whitespaces', noMultipleWhitespaces, {
       output:
         "import {defineMessage} from 'react-intl';defineMessage({defaultMessage: `a\\\\ \\` {placeHolder}`})",
     },
+    {
+      code: "import {defineMessage} from 'react-intl';defineMessage({defaultMessage: `<em>a  b</em>`})",
+      errors: [{message: 'Multiple consecutive whitespaces are not allowed'}],
+      output:
+        "import {defineMessage} from 'react-intl';defineMessage({defaultMessage: `<em>a b</em>`})",
+    },
+    // Multi-line JSX attribute
+    {
+      code: `<FormattedMessage defaultMessage="a\n  b" />`,
+      errors: [{message: 'Multiple consecutive whitespaces are not allowed'}],
+      output: `<FormattedMessage defaultMessage="a b" />`,
+    },
   ],
 })