+++ packages/intl-locale/tests/index.test.ts
@@ -102,4 +102,8 @@ describe('intl-locale', () => {
       )
     })
   })
+  
+  it('has static polyfilled property', function () {
+    expect(Locale.polyfilled).toBe(true)
+  })
 })